Add files via upload
This report discusses the development of the popular board game Chess using the C++ standard template library in an object-oriented framework using Visual Studio 2013.  The decisions made in the game design and implementation are discussed with reference to important features of object-oriented programming.  Improvements that could be made to the game design with both the C++ standard library and beyond are briefly presented.